% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{add_total_graphs_to_results}
\alias{add_total_graphs_to_results}
\title{Add Total Graphs Columns to Results Data Frame}
\usage{
add_total_graphs_to_results(test_set, results)
}
\arguments{
\item{test_set}{The test set used to get the results data frame. The test set
must be created with \code{\link[=make_csafe_cvl_set]{make_csafe_cvl_set()}}.}

\item{results}{A results data frame created with
\code{\link[handwriterRF:compare_writer_profiles]{handwriterRF::compare_writer_profiles()}} and test_set.}
}
\value{
A data frame
}
\description{
The results data frame created by \code{\link[handwriterRF:compare_writer_profiles]{handwriterRF::compare_writer_profiles()}}
does not include total graphs columns. \code{add_total_graphs_to_results()} adds
total_graphs1 and total_graphs2 columns to results to show the number of
graphs in the documents in docname1 and docname2, respectively.
}
\examples{
\dontrun{
test <- make_csafe_cvl_set(all_writers = unique(handwriterRF::test$writer))
results <- handwriterRF::compare_writer_profiles(test$long, test$short, score_only = FALSE)
add_total_graphs_to_results(test, results)
}

}
