% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csafe_prompt_cfr}
\alias{csafe_prompt_cfr}
\title{CSAFE Prompt Cluster Fill Rates Data Frame}
\format{
A data frame with 12824 rows and 46 variables:
\describe{
  \item{docname}{Filename of the handwriting prompt}
  \item{writer}{Writer ID (w0001--w0720)}
  \item{session}{Writing session (s01, s02, s03)}
  \item{prompt}{Writing prompt code (pPHR, pLND, pWOZ)}
  \item{rep}{Repetition number within the given session (r01, r02, r03)}
  \item{total_graphs}{Total number of graphs, or component shapes, into which the handwriting is split}
  \item{cluster1}{Proportion of graphs assigned to cluster1 (0--1)}
  \item{cluster2}{Proportion of graphs assigned to cluster2 (0--1)}
  \item{cluster3}{Proportion of graphs assigned to cluster3 (0--1)}
  \item{cluster4}{Proportion of graphs assigned to cluster4 (0--1)}
  \item{cluster5}{Proportion of graphs assigned to cluster5 (0--1)}
  \item{cluster6}{Proportion of graphs assigned to cluster6 (0--1)}
  \item{cluster7}{Proportion of graphs assigned to cluster7 (0--1)}
  \item{cluster8}{Proportion of graphs assigned to cluster8 (0--1)}
  \item{cluster9}{Proportion of graphs assigned to cluster9 (0--1)}
  \item{cluster10}{Proportion of graphs assigned to cluster10 (0--1)}
  \item{cluster11}{Proportion of graphs assigned to cluster11 (0--1)}
  \item{cluster12}{Proportion of graphs assigned to cluster12 (0--1)}
  \item{cluster13}{Proportion of graphs assigned to cluster13 (0--1)}
  \item{cluster14}{Proportion of graphs assigned to cluster14 (0--1)}
  \item{cluster15}{Proportion of graphs assigned to cluster15 (0--1)}
  \item{cluster16}{Proportion of graphs assigned to cluster16 (0--1)}
  \item{cluster17}{Proportion of graphs assigned to cluster17 (0--1)}
  \item{cluster18}{Proportion of graphs assigned to cluster18 (0--1)}
  \item{cluster19}{Proportion of graphs assigned to cluster19 (0--1)}
  \item{cluster20}{Proportion of graphs assigned to cluster20 (0--1)}
  \item{cluster21}{Proportion of graphs assigned to cluster21 (0--1)}
  \item{cluster22}{Proportion of graphs assigned to cluster22 (0--1)}
  \item{cluster23}{Proportion of graphs assigned to cluster23 (0--1)}
  \item{cluster24}{Proportion of graphs assigned to cluster24 (0--1)}
  \item{cluster25}{Proportion of graphs assigned to cluster25 (0--1)}
  \item{cluster26}{Proportion of graphs assigned to cluster26 (0--1)}
  \item{cluster27}{Proportion of graphs assigned to cluster27 (0--1)}
  \item{cluster28}{Proportion of graphs assigned to cluster28 (0--1)}
  \item{cluster29}{Proportion of graphs assigned to cluster29 (0--1)}
  \item{cluster30}{Proportion of graphs assigned to cluster30 (0--1)}
  \item{cluster31}{Proportion of graphs assigned to cluster31 (0--1)}
  \item{cluster32}{Proportion of graphs assigned to cluster32 (0--1)}
  \item{cluster33}{Proportion of graphs assigned to cluster33 (0--1)}
  \item{cluster34}{Proportion of graphs assigned to cluster34 (0--1)}
  \item{cluster35}{Proportion of graphs assigned to cluster35 (0--1)}
  \item{cluster36}{Proportion of graphs assigned to cluster36 (0--1)}
  \item{cluster37}{Proportion of graphs assigned to cluster37 (0--1)}
  \item{cluster38}{Proportion of graphs assigned to cluster38 (0--1)}
  \item{cluster39}{Proportion of graphs assigned to cluster39 (0--1)}
  \item{cluster40}{Proportion of graphs assigned to cluster40 (0--1)}
}
}
\usage{
csafe_prompt_cfr
}
\description{
A dataset containing the cluster fill rates of 12,824 handwriting prompts
from the CSAFE Handwriting Database. The variables are as follows:
}
\keyword{datasets}
