% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvl_line_cfc}
\alias{cvl_line_cfc}
\title{CVL Line Cluster Fill Counts Data Frame}
\format{
A data frame with 13444 rows and 44 variables:
\describe{
  \item{docname}{Filename of the handwriting prompt}
  \item{writer}{Writer ID (0001--1139)}
  \item{prompt}{Writing prompt code (1, 2, 3, 4, 6, 7, 8)}
  \item{line}{Line number (1--16)}
  \item{1}{Proportion of graphs assigned to cluster1 (0--1)}
  \item{2}{Proportion of graphs assigned to cluster2 (0--1)}
  \item{3}{Proportion of graphs assigned to cluster3 (0--1)}
  \item{4}{Proportion of graphs assigned to cluster4 (0--1)}
  \item{5}{Proportion of graphs assigned to cluster5 (0--1)}
  \item{6}{Proportion of graphs assigned to cluster6 (0--1)}
  \item{7}{Proportion of graphs assigned to cluster7 (0--1)}
  \item{8}{Proportion of graphs assigned to cluster8 (0--1)}
  \item{9}{Proportion of graphs assigned to cluster9 (0--1)}
  \item{10}{Proportion of graphs assigned to cluster10 (0--1)}
  \item{11}{Proportion of graphs assigned to cluster11 (0--1)}
  \item{12}{Proportion of graphs assigned to cluster12 (0--1)}
  \item{13}{Proportion of graphs assigned to cluster13 (0--1)}
  \item{14}{Proportion of graphs assigned to cluster14 (0--1)}
  \item{15}{Proportion of graphs assigned to cluster15 (0--1)}
  \item{16}{Proportion of graphs assigned to cluster16 (0--1)}
  \item{17}{Proportion of graphs assigned to cluster17 (0--1)}
  \item{18}{Proportion of graphs assigned to cluster18 (0--1)}
  \item{19}{Proportion of graphs assigned to cluster19 (0--1)}
  \item{20}{Proportion of graphs assigned to cluster20 (0--1)}
  \item{21}{Proportion of graphs assigned to cluster21 (0--1)}
  \item{22}{Proportion of graphs assigned to cluster22 (0--1)}
  \item{23}{Proportion of graphs assigned to cluster23 (0--1)}
  \item{24}{Proportion of graphs assigned to cluster24 (0--1)}
  \item{25}{Proportion of graphs assigned to cluster25 (0--1)}
  \item{26}{Proportion of graphs assigned to cluster26 (0--1)}
  \item{27}{Proportion of graphs assigned to cluster27 (0--1)}
  \item{28}{Proportion of graphs assigned to cluster28 (0--1)}
  \item{29}{Proportion of graphs assigned to cluster29 (0--1)}
  \item{30}{Proportion of graphs assigned to cluster30 (0--1)}
  \item{31}{Proportion of graphs assigned to cluster31 (0--1)}
  \item{32}{Proportion of graphs assigned to cluster32 (0--1)}
  \item{33}{Proportion of graphs assigned to cluster33 (0--1)}
  \item{34}{Proportion of graphs assigned to cluster34 (0--1)}
  \item{35}{Proportion of graphs assigned to cluster35 (0--1)}
  \item{36}{Proportion of graphs assigned to cluster36 (0--1)}
  \item{37}{Proportion of graphs assigned to cluster37 (0--1)}
  \item{38}{Proportion of graphs assigned to cluster38 (0--1)}
  \item{39}{Proportion of graphs assigned to cluster39 (0--1)}
  \item{40}{Proportion of graphs assigned to cluster40 (0--1)}
}
}
\usage{
cvl_line_cfc
}
\description{
A dataset containing the cluster fill counts of 13,444 lines handwriting from
prompts from the CVL Handwriting Database. The variables are as follows:
}
\keyword{datasets}
