% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{make_csafe_cvl_set}
\alias{make_csafe_cvl_set}
\title{Make CSAFE and CVL Documents Set}
\usage{
make_csafe_cvl_set(
  all_writers,
  csafe_long_prompts = c("pLND", "pWOZ"),
  csafe_num_long = 1,
  csafe_num_short = 2,
  cvl_num_long = 2,
  cvl_num_short = 2,
  cvl_num_lines = 1,
  cvl_drop_German_prompt = TRUE
)
}
\arguments{
\item{csafe_long_prompts}{A vector of CSAFE prompt codes c("pLND", "pWOZ")}

\item{csafe_num_long}{Number of prompts to sample per writer and prompt}

\item{csafe_num_short}{Number of common phrase "pPHR" prompts to sample per
writer}

\item{cvl_num_long}{Number of long documents to sample per writer}

\item{cvl_num_short}{Number of short documents to construct from lines for each
writer}

\item{cvl_num_lines}{Number of lines to use to construct each short document}

\item{writers}{A vector of CSAFE and CVL writer IDs to select from the input
data frames}

\item{drop_German_prompt}{TRUE or FALSE. If TRUE, the prompt written in
German will be dropped from the data frame before randomly selecting
prompts. If FALSE, the German prompt could be randomly selected.}
}
\value{
A list of two data frames
}
\description{
Randomly select short and long CSAFE and CVL documents to construct two data
frames.
}
\examples{
make_csafe_cvl_set(
  all_writers = c("w0001", "w0002", "0023", "0024"),
  csafe_long_prompts = c("pLND", "pWOZ"),
  csafe_num_long = 1,
  csafe_num_short = 2,
  cvl_num_long = 2,
  cvl_num_short = 2,
  cvl_num_lines = 1,
  cvl_drop_German_prompt = TRUE
)

}
