% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{select_cvl_docs}
\alias{select_cvl_docs}
\title{Select CVL Documents}
\usage{
select_cvl_docs(
  writers,
  num,
  drop_German_prompt = TRUE,
  drop_writers_missing_docs = TRUE,
  drop_extra_columns = TRUE
)
}
\arguments{
\item{writers}{A vector of CSAFE writer IDs}

\item{num}{Number of documents to sample per writer}

\item{drop_German_prompt}{TRUE or FALSE. If TRUE, the prompt written in
German will be dropped from the data frame before randomly selecting
prompts. If FALSE, the German prompt could be randomly selected.}

\item{drop_writers_missing_docs}{TRUE or FALSE. Some writers might not have
\code{num} prompts to select. If TRUE, these writers will be dropped from the
data frame. If FALSE, these writers will be included in the data frame.}

\item{drop_extra_columns}{TRUE or FALSE. If TRUE, will keep the docname,
writer and cluster columns and drop all other columns. If FALSE, will keep
all columns.}
}
\value{
A data frame
}
\description{
Randomly select CVL handwriting documents from the cvl_prompt_cfr data frame.
\code{select_cvl_docs} filters the data frame for the specified writer IDs. Then,
the function randomly selects the specified number of documents per writer
and prompt.
}
\examples{
cvl <- select_cvl_docs(
  writers = unique(handwriterRF::train$writer),
  num = 2
)

}
