% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{select_long_short_csafe_docs}
\alias{select_long_short_csafe_docs}
\title{Select Short and Long CSAFE Documents}
\usage{
select_long_short_csafe_docs(
  writers,
  long_prompts,
  num_short,
  num_long,
  drop_extra_columns = TRUE
)
}
\arguments{
\item{writers}{A vector of CSAFE writer IDs}

\item{long_prompts}{A vector of CSAFE prompt codes c("pLND", "pWOZ")}

\item{num_short}{Number of common phrase "pPHR" prompts to sample per writer}

\item{num_long}{Number of prompts to sample per writer and prompt}

\item{drop_extra_columns}{TRUE or FALSE. If TRUE, will keep the docname,
writer and cluster columns and drop all other columns. If FALSE, will keep
all columns.}
}
\value{
A list of two data frames
}
\description{
Randomly select short and long CSAFE documents from the csafe_prompt_cfr data
frame. The common phrase prompts are the only option for short documents.
Choose from London Letter and Wizard of Oz prompts for long documents.
}
\examples{
csafe_train <- select_long_short_csafe_docs(
  writers = unique(handwriterRF::train$writer),
  long_prompts = c("pLND", "pWOZ"),
  num_short = 2,
  num_long = 1
)

}
