% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{select_short_long_csafe_docs}
\alias{select_short_long_csafe_docs}
\title{Select Short and Long CSAFE Documents}
\usage{
select_short_long_csafe_docs(
  df = csafe_prompt_cfr,
  writers,
  long_prompts,
  num_short,
  num_long
)
}
\arguments{
\item{df}{A data frame of CSAFE handwriting documents. The data frame must
contain writer and prompt columns}

\item{writers}{A vector of CSAFE writer IDs}

\item{long_prompts}{A vector of CSAFE prompt codes c("pLND", "pWOZ")}

\item{num_short}{Number of common phrase "pPHR" prompts to sample per writer}

\item{num_long}{Number of prompts to sample per writer and prompt}
}
\value{
A list of two data frames
}
\description{
Randomly select short and long CSAFE documents to construct two data frames.
The common phrase prompts are the only option for short documents. Choose
from London Letter and Wizard of Oz prompts for long documents.
}
\examples{
csafe_train <- select_short_long_csafe_docs(
  df = csafe_prompt_cfr,
  writers = unique(handwriterRF::train$writer),
  long_prompts = c("pLND", "pWOZ"),
  num_short = 2,
  num_long = 1
)
}
