% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{select_short_long_cvl_docs}
\alias{select_short_long_cvl_docs}
\title{Select Short and Long CVL Documents}
\usage{
select_short_long_cvl_docs(
  df = cvl_prompt_cfr,
  lines_df = cvl_line_cfc,
  writers,
  num_short,
  num_long,
  num_lines,
  drop_German_prompt = TRUE
)
}
\arguments{
\item{df}{A data frame of CVL handwriting documents. The data frame must
contain writer and prompt columns}

\item{lines_df}{A data frame of lines of CVL handwriting documents. The
individual lines from the documents in df must be included in lines_df.
Lines from other documents may be included in lines_df as well, but they
will not be used}

\item{writers}{A vector of CVL writer IDs}

\item{num_short}{Number of short documents to construct from lines for each writer}

\item{num_long}{Number of long documents to sample per writer}

\item{num_lines}{Number of lines to use to construct each short document}

\item{drop_German_prompt}{TRUE or FALSE. If TRUE, the prompt written in
German will be dropped from the data frame before randomly selecting
prompts. If FALSE, the German prompt could be randomly selected.}
}
\value{
A list of two data frames
}
\description{
Randomly select CVL documents to construct two data frames - short and long -
with the following steps:
\enumerate{
\item Randomly select num_short + num_long documents
from each writer
\item Assign the first num_short docs selected for each writer
to the short data frame and assign the remaining documents to the long data
frame
\item For each document in the short data frame, randomly select num_lines
from the document and build a pseudo document with \code{\link[=make_pseudo_docs]{make_pseudo_docs()}}
}
}
\examples{
cvl_train <- select_short_long_cvl_docs(
  df = cvl_prompt_cfr,
  lines_df = cvl_line_cfc,
  writers = unique(handwriterRF::train$writer),
  num_short = 2,
  num_long = 2,
  num_lines = 1,
  drop_German_prompt = TRUE
)

}
