% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{make_pseudo_docs}
\alias{make_pseudo_docs}
\title{Make Pseudo Documents from Lines from CVL Documents}
\usage{
make_pseudo_docs(lines_df)
}
\arguments{
\item{lines_df}{A data frame of writer profiles for lines from CVL documents}
}
\value{
A data frame
}
\description{
Make pseudo documents from a data frame of writer profiles for lines from CVL
documents. The writer profiles are summed for each writer to make a pseudo
document.
}
\examples{
cvl <- select_cvl_lines(
  docnames = c("0001-1", "0003-2", "0005-3"),
  lines_df = cvl_line_cfc,
  num_lines = 2
)
docs <- make_pseudo_docs(cvl)

}
