% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{make_pseudo_docs}
\alias{make_pseudo_docs}
\title{Make Pseudo Documents from Lines from CVL Documents}
\usage{
make_pseudo_docs(lines_df, output = "counts")
}
\arguments{
\item{lines_df}{A data frame of cluster fill counts for lines from CVL
documents}

\item{output}{Either "counts" to return cluster fill counts or "rates" to
retun cluster fill rates}
}
\value{
A data frame of cluster fill counts or cluster fill rates
}
\description{
Make pseudo documents from a data frame of writer profiles for lines from CVL
documents. The writer profiles are summed for each writer to make a pseudo
document.
}
\examples{
cvl <- select_cvl_lines(
  docnames = c("0001-1", "0003-2", "0005-3"),
  num_lines = 2
)
counts <- make_pseudo_docs(cvl)
rates <- make_pseudo_docs(cvl, output = "rates")

}
