% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-documents.R
\name{select_csafe_docs}
\alias{select_csafe_docs}
\title{Select CSAFE Documents}
\usage{
select_csafe_docs(df = csafe_prompt_cfr, writers, prompts, num)
}
\arguments{
\item{df}{A data frame of CSAFE handwriting documents. The data frame must
contain writer and prompt columns}

\item{writers}{A vector of CSAFE writer IDs}

\item{prompts}{A vector of CSAFE prompt codes c("pLND", "pPHR", "pWOZ")}

\item{num}{Number of documents to sample per writer and prompt code}
}
\value{
A data frame
}
\description{
Randomly select CSAFE handwriting documents from a data frame.
\code{select_csafe_docs} filters the data frame for the specified writer IDs and
prompts. Then, the function randomly selects the specified number of
documents per writer and prompt.
}
\examples{
csafe <- select_csafe_docs(
  df = csafe_prompt_cfr,
  writers = unique(handwriterRF::train$writer),
  prompts = "pPHR",
  num = 2
)

}
